Sentinel AOT: add SentinelProtectInterceptorHints
